feat: get points by receipt id endpoint
The commit also updates the service model to reflect the range of
values specified by the api contract in the get points response.
The api contract defines points as int64. Java integers are only
32 bit values. The long data type is used instead as it is
a 64 bit value.

+++ src/main/java/com/ctfloyd/receipt/processor/model/receipt/GetReceiptPointsResponse.java
@@ -8,13 +8,13 @@ import java.util.StringJoiner;
  */
 public class GetReceiptPointsResponse {
 
-    private int points;
+    private long points;
 
-    public int getPoints() {
+    public long getPoints() {
         return points;
     }
 
-    public void setPoints(int points) {
+    public void setPoints(long points) {
         this.points = points;
     }
 
@@ -46,7 +46,7 @@ public class GetReceiptPointsResponse {
             getReceiptPointsResponse = new GetReceiptPointsResponse();
         }
 
-        public Builder withPoints(int points) {
+        public Builder withPoints(long points) {
             getReceiptPointsResponse.setPoints(points);
             return this;
         }
